some refactor for custom validator, add missing unit test for util funcs
